backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
nltk
sentence-transformers
pinecone
python-dotenv
